feat: add `searched tools page`
