+++ src/components/common/index.ts
@@ -3,5 +3,6 @@ import Configurations from "./Configurations";
 import Main from "./Main";
 import MainItem from "./MainItem";
 import TextField from "./TextField";
+import ToolCardGrid from "./ToolCardGrid";
 
-export { CodeEditorHalfLoading, Configurations, Main, MainItem, TextField };
+export { CodeEditorHalfLoading, Configurations, Main, MainItem, TextField, ToolCardGrid };
