+++ src/components/pages/search/index.ts
@@ -1,0 +1,3 @@
+import Content from "./Content";
+
+export { Content };
