+++ src/libs/$path.ts
@@ -21,6 +21,9 @@ export const pagesPath = {
       }),
     },
   },
+  search: {
+    $url: (url?: { hash?: string }) => ({ pathname: "/search" as const, hash: url?.hash }),
+  },
   $url: (url?: { hash?: string }) => ({ pathname: "/" as const, hash: url?.hash }),
 };
