+++ src/pages/search.tsx
@@ -1,0 +1,7 @@
+import type { NextPage } from "next";
+
+import { Content } from "@/components/pages/search";
+
+const Page: NextPage = Content;
+
+export default Page;
